% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_attribute,Dataset-method}
\alias{add_attribute,Dataset-method}
\title{Function to add a new attribute to the dataset}
\usage{
\S4method{add_attribute}{Dataset}(object, key, att)
}
\arguments{
\item{key}{Name of the attribute to be added}

\item{att}{Vector containing the attribute data}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function adds the specified attribute to the dataset
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple")))
Weight_data <- c(1,4,2)
MyDataset <- add_attribute(MyDataset,"Weight",Weight_data)
}
