% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{class}
\name{Attribute-class}
\alias{Attribute-class}
\title{An S4 class to represent an Attribute}
\description{
An S4 class to represent an Attribute
}
\section{Slots}{

\describe{
\item{\code{value}}{Vector containing the attribute data}

\item{\code{type}}{Type of the attribute}
}}

