% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attribute}
\alias{attribute}
\title{Basic constructor of the Attribute class}
\usage{
attribute(data)
}
\arguments{
\item{data}{Vector containing the attribute data}
}
\value{
An object of class \code{\linkS4class{Attribute}} that represents the attribute
}
\description{
This function creates an object of class \code{\linkS4class{Attribute}}
}
