% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{correlation1,Dataset-method}
\alias{correlation1,Dataset-method}
\title{Function to compute the correlation between two numerical attributes}
\usage{
\S4method{correlation1}{Dataset}(object, att_A, att_B, method = "pearson")
}
\arguments{
\item{att_A}{Name of the first numerical attribute}

\item{att_B}{Name of the second numerical attribute}

\item{method}{Correlation measure to be used: pearson, spearman, kendall (\code{"pearson"} by default)}
}
\value{
The computed correlation
}
\description{
This function computes the correlation between two numerical attributes
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,23,2,1,4),"Quantity"=c(1,2,1,1,3)))
cor_p <- correlation1(MyDataset,"Weight","Quantity",method="pearson")
cor_s <- correlation1(MyDataset,"Weight","Quantity",method="spearman")
cor_k <- correlation1(MyDataset,"Weight","Quantity",method="kendall")
}
