% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{correlation_att,Dataset-method}
\alias{correlation_att,Dataset-method}
\title{Function to compute the correlation between all the numerical attributes}
\usage{
\S4method{correlation_att}{Dataset}(object, method = "pearson")
}
\arguments{
\item{method}{Correlation measure to be used: pearson, spearman, kendall (\code{"pearson"} by default)}
}
\value{
A matrix containing the computed correlations
}
\description{
This function computes the correlation between all the numerical attributes
}
\examples{
MyDataset <- dataset(list("Code"=c(1,11,2,43,12),"Weight"=c(1,23,2,1,4),"Quantity"=c(1,2,1,1,3)))
cors_p <- correlation_att(MyDataset,method="pearson")
cors_s <- correlation_att(MyDataset,method="spearman")
cors_k <- correlation_att(MyDataset,method="kendall")
}
