% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\title{An S4 class to represent a Dataset}
\description{
An S4 class to represent a Dataset
}
\section{Slots}{

\describe{
\item{\code{data}}{List containing the attributes of the dataset (class \code{\linkS4class{Attribute}})}

\item{\code{class_att}}{Name of the class variable}

\item{\code{length}}{Number of instances in the dataset}
}}

