% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Basic constructor of the Dataset class}
\usage{
dataset(data, c = "")
}
\arguments{
\item{data}{List or dataframe containing the data}

\item{c}{Name of the class variable (\code{""} by default)}
}
\value{
An object of class \code{\linkS4class{Dataset}} that represents the dataset
}
\description{
This function creates an object of class \code{\linkS4class{Dataset}}
}
