% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{discretize1,Dataset-method}
\alias{discretize1,Dataset-method}
\title{Function to discretize a numerical attribute}
\usage{
\S4method{discretize1}{Dataset}(object, att, method, num_bins = NULL, cut_points = NULL)
}
\arguments{
\item{att}{Name of the numerical attribute to be discretized}

\item{method}{Discretization method to be used: frequency, width, custom}

\item{num_bins}{Number of intervals (only for frequency and width methods, \code{NULL} by default)}

\item{cut_points}{Vector containing the interval cut points (only for custom method, \code{NULL} by default)}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function discretizes the specified numerical attribute
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,12,10,5)))
MyDataset1 <- discretize1(MyDataset,"Weight",method="frequency",num_bins=2)
MyDataset2 <- discretize1(MyDataset,"Weight",method="width",num_bins=2)
MyDataset3 <- discretize1(MyDataset,"Weight",method="custom",cut_points=c(4,11))
}
