% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{discretize_att,Dataset-method}
\alias{discretize_att,Dataset-method}
\title{Function to discretize all the numerical attributes}
\usage{
\S4method{discretize_att}{Dataset}(object, method, num_bins)
}
\arguments{
\item{method}{Discretization method to be used: frequency, width}

\item{num_bins}{Number of intervals}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function discretizes all the numerical attributes
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,12,10,5),"Quantity"=c(10,4,5,2,7,56)))
MyDataset1 <- discretize_att(MyDataset,method="frequency",num_bins=2)
MyDataset2 <- discretize_att(MyDataset,method="width",num_bins=2)
}
