% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{entropy1,Dataset-method}
\alias{entropy1,Dataset-method}
\title{Function to compute the entropy of a non-numerical attribute}
\usage{
\S4method{entropy1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the non-numerical attribute}
}
\value{
The computed entropy
}
\description{
This function computes the entropy of a non-numerical attribute
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon")))
entropy <- entropy1(MyDataset,"Fruit")
}
