% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{entropy_att,Dataset-method}
\alias{entropy_att,Dataset-method}
\title{Function to compute the entropy of all the non-numerical attributes}
\usage{
\S4method{entropy_att}{Dataset}(object)
}
\value{
A list containing the computed entropies
}
\description{
This function computes the entropy of all the non-numerical attributes
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon"),"Expensive"=c(TRUE,TRUE,TRUE,FALSE)))
entropies <- entropy_att(MyDataset)
}
