% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from_csv}
\alias{from_csv}
\title{Function to read the dataset from a csv file}
\usage{
from_csv(f, cat = 5, header = TRUE, sep = ",", c = "")
}
\arguments{
\item{f}{Path of the csv file}

\item{cat}{Minimum number of possible values for a character attribute. Any character attribute with less than
cat possible values is converted to factor (\code{5} by default)}

\item{header}{Boolean that indicates whether the csv file contains a header or not (\code{TRUE} by default)}

\item{sep}{Character that is used to separate the data in the csv file (\code{","} by default)}

\item{c}{Name of the class variable (\code{""} by default)}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function reads the dataset from a csv file
}
