% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mean1,Dataset-method}
\alias{mean1,Dataset-method}
\title{Function to compute the mean of a numerical attribute}
\usage{
\S4method{mean1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute}
}
\value{
The computed mean
}
\description{
This function computes the mean of a numerical attribute
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,4,5,2)))
mean <- mean1(MyDataset,"Weight")
}
