% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mean_att,Dataset-method}
\alias{mean_att,Dataset-method}
\title{Function to compute the mean of all the numerical attributes}
\usage{
\S4method{mean_att}{Dataset}(object)
}
\value{
A list containing the computed means
}
\description{
This function computes the mean of all the numerical attributes
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,4,5),"Quantity"=c(34,1,23,2,1)))
means <- mean_att(MyDataset)
}
