% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{median1,Dataset-method}
\alias{median1,Dataset-method}
\title{Function to compute the median of a numerical attribute}
\usage{
\S4method{median1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute}
}
\value{
The computed median
}
\description{
This function computes the median of a numerical attribute
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,4,5,2)))
median <- median1(MyDataset,"Weight")
}
