% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mode1,Dataset-method}
\alias{mode1,Dataset-method}
\title{Function to compute the mode of a non-numerical attribute}
\usage{
\S4method{mode1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the non-numerical attribute}
}
\value{
The computed mode
}
\description{
This function computes the mode of a non-numerical attribute.
If there is a tie, the value that appears first is returned
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon")))
mode <- mode1(MyDataset,"Fruit")
}
