% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mode_att,Dataset-method}
\alias{mode_att,Dataset-method}
\title{Function to compute the mode of all the non-numerical attributes}
\usage{
\S4method{mode_att}{Dataset}(object)
}
\value{
A list containing the computed modes
}
\description{
This function computes the mode of all the non-numerical attributes.
If there is a tie, the value that appears first is returned
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon"),"Expensive"=c(TRUE,TRUE,TRUE,FALSE)))
modes <- mode_att(MyDataset)
}
