% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{norm_mutual_info1,Dataset-method}
\alias{norm_mutual_info1,Dataset-method}
\title{Function to compute the normalized mutual information between two non-numerical attributes}
\usage{
\S4method{norm_mutual_info1}{Dataset}(object, att_A, att_B)
}
\arguments{
\item{att_A}{Name of the first non-numerical attribute}

\item{att_B}{Name of the second non-numerical attribute}
}
\value{
The computed normalized mutual information
}
\description{
This function computes the normalized mutual information between two non-numerical attributes
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon"),"Expensive"=c(TRUE,TRUE,FALSE,FALSE)))
nmi <- norm_mutual_info1(MyDataset,"Fruit","Expensive")
}
