% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{normalize1,Dataset-method}
\alias{normalize1,Dataset-method}
\title{Function to normalize a numerical attribute}
\usage{
\S4method{normalize1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute to be normalized}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function normalizes the specified numerical attribute
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,12,10,5)))
MyDataset <- normalize1(MyDataset,"Weight")
}
