% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_norm_mutual_info}
\alias{plot_norm_mutual_info}
\title{Function to plot the normalized mutual information matrix of the dataset}
\usage{
plot_norm_mutual_info(object)
}
\description{
This function plots the normalized mutual information matrix of the dataset
}
\details{
This function requires two additional packages, ggplot2 and reshape2
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple","Watermelon"),"Shop"=c("A","B","A","C"),"Expensive"=c(TRUE,TRUE,FALSE,FALSE)))
plot_norm_mutual_info(MyDataset)
}
