% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_roc}
\alias{plot_roc}
\title{Function to plot the ROC curve of a numerical attribute (when the class is boolean)}
\usage{
plot_roc(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute}
}
\description{
This function plots the ROC curve of a numerical attribute (when the class is boolean)
}
\details{
This function requires the additional package ggplot2
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,23,2,1,4),"Expensive"=c(TRUE,FALSE,TRUE,FALSE,FALSE)),c="Expensive")
plot_roc(MyDataset,"Weight")
}
