% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{print_data,Attribute-method}
\alias{print_data,Attribute-method}
\title{Function to print the attribute data}
\usage{
\S4method{print_data}{Attribute}(object)
}
\description{
This function prints the attribute data
}
\examples{
MyAttribute <- attribute(c(1,2,3,1,2))
print_data(MyAttribute)
}
