% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{print_dataset,Dataset-method}
\alias{print_dataset,Dataset-method}
\title{Function to print the dataset data}
\usage{
\S4method{print_dataset}{Dataset}(object)
}
\description{
This function prints the dataset data
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple")))
print_dataset(MyDataset)
}
