% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{remove_attribute,Dataset-method}
\alias{remove_attribute,Dataset-method}
\title{Function to remove an attribute from the dataset}
\usage{
\S4method{remove_attribute}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the attribute to be removed}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function removes the specified attribute from the dataset
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple")))
MyDataset <- remove_attribute(MyDataset,"Fruit")
}
