% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{roc_auc1,Dataset-method}
\alias{roc_auc1,Dataset-method}
\title{Function to compute the AUC of a numerical attribute (when the class is boolean)}
\usage{
\S4method{roc_auc1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute}
}
\value{
The computed AUC
}
\description{
This function computes the AUC of a numerical attribute (when the class is boolean)
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,23,2,1,4),"Expensive"=c(TRUE,FALSE,TRUE,FALSE,FALSE)),c="Expensive")
auc <- roc_auc1(MyDataset,"Weight")
}
