% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{roc_auc_att,Dataset-method}
\alias{roc_auc_att,Dataset-method}
\title{Function to compute the AUC of all the numerical attributes (when the class is boolean)}
\usage{
\S4method{roc_auc_att}{Dataset}(object)
}
\value{
A list containing the computed AUCs
}
\description{
This function computes the AUC of all the numerical attributes (when the class is boolean)
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,23,2,1,4),"Quantity"=c(1,2,1,1,3),"Expensive"=c(TRUE,FALSE,TRUE,FALSE,FALSE)),c="Expensive")
aucs <- roc_auc_att(MyDataset)
}
