% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{standarize1,Dataset-method}
\alias{standarize1,Dataset-method}
\title{Function to standarize a numerical attribute}
\usage{
\S4method{standarize1}{Dataset}(object, att)
}
\arguments{
\item{att}{Name of the numerical attribute to be standarized}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function standarizes the specified numerical attribute
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,12,10,5)))
MyDataset <- standarize1(MyDataset,"Weight")
}
