% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{standarize_att,Dataset-method}
\alias{standarize_att,Dataset-method}
\title{Function to standarize all the numerical attributes}
\usage{
\S4method{standarize_att}{Dataset}(object)
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function standarizes all the numerical attributes
}
\examples{
MyDataset <- dataset(list("Weight"=c(1,3,2,12,10,5),"Quantity"=c(10,23,2,1,0,2)))
MyDataset <- standarize_att(MyDataset)
}
