% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{to_csv,Dataset-method}
\alias{to_csv,Dataset-method}
\title{Function to write the dataset to a csv file}
\usage{
\S4method{to_csv}{Dataset}(object, f, header = TRUE, sep = ",")
}
\arguments{
\item{f}{Path of the csv file}

\item{header}{Boolean that indicates whether the header should be included or not (\code{TRUE} by default)}

\item{sep}{Character that is used to separate the data in the csv format (\code{","} by default)}
}
\description{
This function writes the dataset to a csv file
}
