% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{to_dataframe,Dataset-method}
\alias{to_dataframe,Dataset-method}
\title{Function to convert the dataset to a dataframe}
\usage{
\S4method{to_dataframe}{Dataset}(object)
}
\value{
A dataframe object containing the dataset data
}
\description{
This function converts the dataset to a dataframe object
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple")))
MyDataframe <- to_dataframe(MyDataset)
}
