% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{to_factor,Dataset-method}
\alias{to_factor,Dataset-method}
\title{Function to convert a non-numerical attribute to factor}
\usage{
\S4method{to_factor}{Dataset}(object, att, levels = NULL)
}
\arguments{
\item{levels}{Vector containing the levels of the factor attribute (\code{NULL} by default)}

\item{key}{Name of the non-numerical attribute to be converted}
}
\value{
An object of class \code{\linkS4class{Dataset}}
}
\description{
This function converts the specified non-numerical attribute to factor
}
\examples{
MyDataset <- dataset(list("Fruit"=c("Apple","Orange","Apple")))
MyDataset <- to_factor(MyDataset,"Fruit",levels=c("Apple","Orange"))
}
